feat(stella-pipeline): a shell is a witness runner, so runner-less workspaces can author (#2064)
The witness stage degraded on every Terminal-Bench task with 'no supported
test runner is available in this workspace': RUNNER_VOCABULARY listed only
language toolchains, and TB containers are bare system images — nginx
config, git surgery, MIPS interpreters. Measured: witness_authored was 1 of
51 trials across the whole arenabench corpus, and 0 across every cycle run
on 2026-08-06/07.

sh joins the vocabulary as the floor: it is present in every container the
agent already executes in, and the witness contract — fails on the old
code, passes on the new — is exactly a script's exit code. The probe is
sh -c 'exit 0' (sh --version is non-portable; dash and busybox reject it),
carved out in the version-style probe invariant with its reason.

Acceptance is deliberately narrower than the issue sketched: only
'sh <script>.sh' parses, never 'sh -c <inline>'. An inline program rides as
one opaque argv word, so the per-argument confinement checks cannot see a
path inside it — and an invocation that names no artifact could never be
tamper-excluded. The script path gets the same absolute/.. escape checks as
every other runner's arguments, validate_witness_invocation requires the
invocation to pin exactly the authored artifact, and is_witness_test_path
requires the name to declare intent (witness_*/test_*/tests dir) so an
arbitrary deploy.sh is not a witness artifact.

The author is TOLD the shape when the probe found only a shell — a model
asked for 'a test' in a runner-less container writes a pytest file
otherwise — via a dedicated section in witness_prompt, derived from the
probed availability set (#1539's discipline), plus a system-prompt line.
The vacuity guard needs no sh-specific work: the baseline-passes rejection
in witness_stage never branches on the runner.

witness.rs's inline test module moves to witness/tests.rs verbatim (the
#1914 extraction pattern) — the file was 1438 lines and four lines of
growth from the 1500 ceiling was not room to add an arm.

Closes #2064
